Code to plot timeseries as anomalies, allowing for class column.
Code to read in raw datafile and plot anomalies, ie. each data point is plotted as (value - mean value across all years for that month). Output pdf contains individual anomaly timeseries for each plot (1 per page). This is a version of PlotAnomalies.R, altered to allow for a 'class' column, for use with dendrometer data.